:white_check_mark:

+++ crates/core_simd/tests/isize_ops.rs
@@ -1,5 +1,6 @@
+
 #![feature(portable_simd)]
 
 #[macro_use]
 mod ops_macros;
-impl_signed_tests! { i64 }
+impl_signed_tests! { isize }
